+++ src/Avalonia.Input/DragDrop/Raw/RawDragEventType.cs
@@ -1,4 +1,4 @@
-﻿namespace Avalonia.Controls.DragDrop.Raw
+﻿namespace Avalonia.Input.DragDrop.Raw
 {
     public enum RawDragEventType
     {
