+++ src/OSX/Avalonia.MonoMac/DragSource.cs
@@ -10,8 +10,8 @@ using System.Runtime.InteropServices;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Threading.Tasks;
 using Avalonia.Controls;
-using Avalonia.Controls.DragDrop;
-using Avalonia.Controls.Platform;
+using Avalonia.Input.DragDrop;
+using Avalonia.Input.Platform;
 using Avalonia.Input;
 using Avalonia.Input.Raw;
 using MonoMac;
