+++ src/OSX/Avalonia.MonoMac/DraggingInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Avalonia.Controls.DragDrop;
+using Avalonia.Input.DragDrop;
 using MonoMac.AppKit;
 using MonoMac.Foundation;
